requirements.txt
sounddevice
numpy
pyttsx3
# Optional:
google-cloud-speech
vosk
